Refactor tests
- Convert into integration tests
- Additional CI checks
