+++ .gitignore
@@ -3,3 +3,4 @@ Cargo.lock
 scripts/tmp
 scripts/*.txt
 scripts/*.rs
+bench_data/*
